fix iNode data type

+++ pkg/dataplane/types.go
@@ -186,7 +186,7 @@ type GetObjectInput struct {
 
 type GetObjectByInodeInput struct {
 	DataPlaneInput
-	InodeNumber uint32
+	InodeNumber uint64
 	Slice       int
 	CTimeSecs   int64
 	CTimeNanos  int64
